side panel now reusable component
